Fixed windowSize parameter
I fixed the windowSize parameter so it does now change the window size
of the audioNode during operation. It took me a while to guess that I
needed to add a disconnectAll() function to the audio_nodes files so I
could both disconnect and disable the node before reconfiguring it.
I also renamed the superficial shift parameters I was using to "search"
names instead to more closely match the search function of an
oscilloscope since that's really all those are doing.