Fixed a few bugs with changing the windowSize
Changed code in audio_nodes.cpp so the setup function only creates a new
input node if one didn't previously exist. Also, fixed it so it didn't
keep adding the input name to the window title. Not sure if I need all
the stuff I added to the disconnectAll function of the audio_nodes.cpp
file though...